Replicate native API for WASM
