+++ index.js
@@ -1,0 +1,4 @@
+if (typeof localStorage === "undefined" || localStorage === null)
+  module.exports = require("./native/index")
+else
+  module.exports = require("./web/index")
